refactor(parser): CalculatorProcess에서 parsing 메서드를 분리한다
Parsing 관련 메서드를 모듈화(Parser.js)하여 관심사를 분리한다.
