+++ src/CalculatorProcess.js
@@ -1,8 +1,7 @@
 import { Console } from '@woowacourse/mission-utils';
+import Parser from './Parser.js';
 
 class CalculatorProcess {
-  inputNums = [];
-
   inputString = '';
 
   async getString() {
@@ -15,44 +14,9 @@ class CalculatorProcess {
     this.inputString = initInput;
   }
 
-  handleDefaultDelimiter() {
-    if (/^\d+([,:]\d+)*$/.test(this.inputString)) {
-      const parsedNums = this.inputString.match(/\d+/g);
-      this.inputNums = parsedNums.map(Number);
-      return true;
-    }
-    return false;
-  }
-
-  handleCustomDelimiter() {
-    const customDelimiterMatch = this.inputString.match(/^\/\/(.)\\n(.*)$/);
-
-    if (customDelimiterMatch) {
-      const customDelimiter = this.escapeRegExp(customDelimiterMatch[1]);
-      const delimiterRegex = new RegExp(`[${customDelimiter},:]`);
-      const validationRegex = new RegExp(`^\\d+([${customDelimiter},:]\\d+)*$`);
-
-      if (validationRegex.test(customDelimiterMatch[2])) {
-        this.inputString = customDelimiterMatch[2];
-      } else {
-        throw new Error('[ERROR] 계산을 위한 문자열의 형식이 잘못 되었습니다.');
-      }
-
-      const parsedNums = this.inputString.split(delimiterRegex).map(Number);
-      this.inputNums = parsedNums;
-    } else {
-      throw new Error('[ERROR] 커스텀 구분자 할당 형식이 잘못 되었습니다.');
-    }
-  }
-
-  escapeRegExp(string) {
-    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  }
-
   doParsing() {
-    if (!this.handleDefaultDelimiter()) {
-      this.handleCustomDelimiter();
-    }
+    const parser = new Parser(this.inputString);
+    this.inputNums = parser.parse();
   }
 
   operatingNums() {
